fix: fix getFrontZIndex logic
